Changed price and balance from int to bigdecimal

+++ java/database/schema.sql
@@ -19,7 +19,7 @@ CREATE SEQUENCE seq_account_id
 CREATE TABLE account (
 	account_id int NOT NULL DEFAULT nextval('seq_account_id'),
 	user_id int NOT NULL,
-	balance int NOT NULL,
+	balance decimal(13, 2) NOT NULL,
 	CONSTRAINT PK_account PRIMARY KEY (account_id),
 	CONSTRAINT FK_account_user FOREIGN KEY (user_id) REFERENCES users (user_id)
 );
@@ -32,7 +32,7 @@ CREATE TABLE property (
     bedrooms int NOT NULL,
     bathrooms int NOT NULL,
     property_type varchar(20) NOT NULL,
-    price int  NOT NULL,
+    price decimal(13, 2) NOT NULL,
     address varchar(100) NOT NULL,
     availability boolean NOT NULL,
     property_desc varchar(2000) NOT NULL,
